Move pokemon domain models to core module
