--- app/src/main/java/mx/yellowme/youst/domain/PokemonTypeSelection.kt
@@ -1,6 +1,0 @@
-package mx.yellowme.youst.domain;
-
-/**
- * TODO: Add docs
- */
-class PokemonTypeSelection(var type: PokemonType?, var isChecked: Boolean)
