+++ app/src/main/java/mx/yellowme/youst/pokemon/index/PokemonListAdapter.kt
@@ -5,14 +5,14 @@ import android.view.ViewGroup
 import mx.yellowme.youst.R
 import mx.yellowme.youst.core.hooks.recycler.ItemListener
 import mx.yellowme.youst.core.hooks.recycler.SimpleRecyclerAdapter
-import mx.yellowme.youst.domain.Pokemon
+import mx.yellowme.youst.core.domain.Pokemon
 
 /**
  * MainModel adapter for displaying a list.
  */
 class PokemonListAdapter(
-    items: MutableList<Pokemon>,
-    itemListener: ItemListener<Pokemon>
+        items: MutableList<Pokemon>,
+        itemListener: ItemListener<Pokemon>
 ) : SimpleRecyclerAdapter<Pokemon, PokemonListViewHolder>(items, itemListener) {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): PokemonListViewHolder {
