+++ app/src/main/java/mx/yellowme/youst/pokemon/index/PokemonListViewHolder.kt
@@ -9,7 +9,7 @@ import com.squareup.picasso.Picasso
 import mx.yellowme.youst.R
 import mx.yellowme.youst.core.hooks.recycler.ItemListener
 import mx.yellowme.youst.core.hooks.recycler.RecyclerViewHolderDecorator
-import mx.yellowme.youst.domain.Pokemon
+import mx.yellowme.youst.core.domain.Pokemon
 
 /**
  * Reusable view portion.
