+++ core/src/main/java/mx/yellowme/youst/core/domain/Pokemon.kt
@@ -1,4 +1,4 @@
-package mx.yellowme.youst.domain;
+package mx.yellowme.youst.core.domain;
 
 /**
  * Main application domain model.
