+++ core/src/main/java/mx/yellowme/youst/core/domain/PokemonType.kt
@@ -1,8 +1,8 @@
-package mx.yellowme.youst.domain;
+package mx.yellowme.youst.core.domain;
 
 import android.content.Context
 import androidx.core.content.ContextCompat
-import mx.yellowme.youst.R
+import mx.yellowme.youst.core.R
 
 /**
  * TODO: Add docs
