+++ core/src/main/java/mx/yellowme/youst/core/domain/PokemonTypeSelection.kt
@@ -1,0 +1,6 @@
+package mx.yellowme.youst.core.domain
+
+/**
+ * TODO: Add docs
+ */
+class PokemonTypeSelection(var type: PokemonType?, var isChecked: Boolean)
